% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calc_helpers.R
\name{agg_by_deciles}
\alias{agg_by_deciles}
\title{aggregate cariables of a single simulaiton by decile}
\usage{
agg_by_deciles(
  dta,
  policy_name = NA_character_,
  dec_by,
  dec_vars,
  n_dec = 10,
  wt_var = NULL,
  get_var_fn = get_var_nm,
  dec_min_level = NULL,
  dec_max_level = NULL,
  ...
)
}
\description{
aggregate cariables of a single simulaiton by decile
}
