% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calc_helpers.R
\name{calc_agg_by}
\alias{calc_agg_by}
\title{Aggregates all relevant CEQ results by variable.}
\usage{
calc_agg_by(dta, vars, by_var, wt_var = get_wt_nm())
}
\arguments{
\item{dta}{data frame with the simulation results.}

\item{by_var}{character name of the grouping variable}

\item{wt_var}{character name of a single variable that should be used
as a weighting variable in calculation. Default is get_wt_nm()$var.
If wt_var is not provided or the one provided is not present in `dta`,
warning will be prompted.}
}
\description{
Aggregates all relevant CEQ results by variable.
}
