% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gini_poverty.R
\name{calc_pov_fgt}
\alias{calc_pov_fgt}
\title{Compute Theil index from a numeric vector}
\usage{
calc_pov_fgt(x, pl, alpha = 0, w = rep(1, length(x)), na.rm = TRUE, ...)
}
\arguments{
\item{x, w}{numeric vectors of values and population weights of equal length}

\item{pl}{poverty line}

\item{alpha}{Foster-Greer-Thorbecke (FGT) povert parameter}

\item{na.rm}{logical `TRUE` bu default. To omit NA values}
}
\description{
Compute Theil index from a numeric vector
}
