% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gini_poverty.R
\name{calc_theil}
\alias{calc_theil}
\title{Compute Theil index from a numeric vector}
\usage{
calc_theil(x, w = rep(1, length(x)), na.rm = TRUE, drop_zero_and_less = TRUE)
}
\arguments{
\item{x, w}{numeric vectors of values and population weights of equal length}

\item{na.rm}{logical `TRUE` bu default. To omit NA values}

\item{drop_zero_and_less}{logical `TRUE` bu default to drop negarive values
and zeros}
}
\description{
Compute Theil index from a numeric vector
}
