% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEQ_run.R
\name{CEQ_run}
\alias{CEQ_run}
\alias{CEQ_run_example}
\title{Run the Shiny Application}
\usage{
CEQ_run(
  onStart = NULL,
  options = list(),
  enableBookmarking = NULL,
  uiPattern = "/",
  inputs_str,
  presim,
  baseline_dta = reactive(tibble(var = "Baseline from `CEQ_run`")),
  n_policy = c(1, 2, 1),
  n_policy_type = c("numericInline", "numeric", "slider", "dropdown", "none"),
  server_fn = CEQ_server,
  ui_fn = CEQ_ui,
  inp_str_fn = gen_inp_str,
  ui_gen_fn = gen_tabinp_ui,
  ceq_fn = function(inps, presim) {
     tibble(var = "Results from `CEQ_run`")
 },
  info_page_md = NULL,
  info_page_size = "l",
  fn_sim_srvr = make_run_sim_server(),
  fn_add_missing_inp = NULL,
  fn_ceq_sim = NULL,
  fn_ceq_pre_postsim = NULL,
  fn_postsim_srvr = mod_generic_run_postsim_server,
  fn_ceq_postsim = function(x, ...) x,
  fn_res_disp_srvr = fn_results_display_server_dummy,
  ...
)

CEQ_run_example(...)
}
\arguments{
\item{inputs_str}{UI inputs structure that is read with `load_input_xlsx` or
another custom function.}

\item{presim}{reactive with the list of pre-simulation parameters relevant
to the analysis}

\item{n_policy}{vector of 3 numeric values c(1,2,1) by default.
The first and the second are minimum and maximum number of policy choices.
The numeric is the initialized number of policy choices.}

\item{n_policy_type}{character that defines types of the the changer of the
the number of policy scenarios. Could be one of "numericInline" (default),
"numeric", "slider", "dropdown", and "none".}

\item{...}{not used}
}
\description{
Run the Shiny Application
}
\section{Functions}{
\itemize{
\item \code{CEQ_run_example()}: Run example shiny app with an elementary CEQ analysis

}}
