% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEQ_ui.R
\name{CEQ_ui}
\alias{CEQ_ui}
\title{The application User-Interface}
\usage{
CEQ_ui(
  request,
  theme_fn = function() {
     bslib::bs_theme(version = 4, bootswatch = "flatly",
    `enable-rounded` = TRUE)
 },
  style_logo_position = NULL,
  inp_nav_width = NULL,
  fn_results_ui = fn_results_ui_dummy,
  ...
)
}
\arguments{
\item{request}{Internal parameter for `{shiny}`.
DO NOT REMOVE.}
}
\description{
The application User-Interface
}
