% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_microsim.R
\name{create_microsim}
\alias{create_microsim}
\title{Create a directory with key apps' files and `app.R` needed for development
and deployment of the microsimulation Shiny app}
\usage{
create_microsim(
  path,
  open = TRUE,
  app_name = basename(path),
  newSession = FALSE
)
}
\arguments{
\item{path}{full path to the place where the app's folder must be along with
the not yet existing app folder name. The folder will be created automatically.}

\item{open}{logical, if the app's folder should be opened as an RStudio project.}

\item{app_name}{optional. name of folder where the app should be.}

\item{newSession}{Boolean (default FALSE); should the project be opened in a
new session, or #' should the current RStudio session switch to that project?}
}
\value{
The return value, if any, from executing the function.
}
\description{
Copies a template app folder structure to the location/name
  specified. Creates `app.R` as well as the RStudio project to simplify
  the  workflow.
}
