% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEQ_ui.R
\name{gen_ceq_ui}
\alias{gen_ceq_ui}
\title{Generate a local CEQ UI function}
\usage{
gen_ceq_ui(
  fn_results_ui = fn_results_ui_dummy,
  theme_fn = function() shinythemes::shinytheme("cerulean"),
  style_logo_position = NULL,
  inp_nav_width = NULL,
  ...
)
}
\description{
Generate a local CEQ UI function
}
