% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tab_input_ui.R
\name{gen_one_inp_table}
\alias{gen_one_inp_table}
\title{Generate a table with inputs instead of inputs with lables}
\usage{
gen_one_inp_table(inp_ui_str, inp_table_str_one)
}
\description{
Generate a table with inputs instead of inputs with lables
}
