% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tab_input_ui.R
\name{gen_tabinp_ui_front}
\alias{gen_tabinp_ui_front}
\title{Generates a function based on `gen_tabinp_ui` for generating tabs UI}
\usage{
gen_tabinp_ui_front(
  inp_tab_str = NULL,
  inp_table_str = NULL,
  scroll_panel_style =
    "max-height: calc(100vh - 12rem); overflow-y: scroll; overflow-x: hidden;"
)
}
\arguments{
\item{inp_tab_str}{is a tibble with the structure of tabs in the app.
Requires two columns: `tab_name` and `group_order`. The first one indicates
the name of the tab. The second one indicates the order of a well-group for
for including under this tab.}
}
\description{
Generates a function based on `gen_tabinp_ui` for generating tabs UI
}
