% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_input_prep_key.R
\name{get_all_inps}
\alias{get_all_inps}
\title{Return all inputs from the raw inputs structure}
\usage{
get_all_inps(dta)
}
\arguments{
\item{data}{dataframe with the raw inputs structure}
}
\description{
Return all inputs from the raw inputs structure
}
