% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gini_poverty.R
\name{get_dta_pov}
\alias{get_dta_pov}
\title{Calculate poverty  for a single simulation at the national poverty line}
\usage{
get_dta_pov(
  dta,
  policy_name = NULL,
  poverty_line_var = NULL,
  poverty_line_value = NULL,
  income_vars_tbl = NULL,
  wt_var = NULL,
  para_names = tibble(parameter = c("rate", "headcount", "gap", "severity"), label =
    c("Poverty rate", "Poverty headcount", "Poverty gap", "Poverty severity") \%>\%
    factor(., levels = .)),
  ...
)
}
\arguments{
\item{dta}{data frame with the simulation results.}

\item{income_vars_tbl}{a data frame with two character columns: `var` and
`var_title`. `var` is the column with the variable names returned from
the data set. `var_title` is the variable name as we want it to appear
in the resulting table. See for example results of the functions
`get_inc_nm()` or `get_var_nm()`}

\item{wt_var}{character name of a single variable that should be used
as a weighting variable in calculation. Default is get_wt_nm()$var.
If wt_var is not provided or the one provided is not present in `dta`,
warning will be prompted.}

\item{poverty_line}{numeric, indicating the poverty line in Indonesian Rupiah
per person per day.}
}
\description{
Calculate poverty  for a single simulation at the national poverty line
}
