% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_sim_helpers.R
\name{make_get_inp_fn}
\alias{make_get_inp_fn}
\title{Return a function that extracts inputs frmo the inputs list do not returning NAs}
\usage{
make_get_inp_fn(actual_inps)
}
\arguments{
\item{actual_inps}{is the inputs list with the inputs}
}
\description{
Return a function that extracts inputs frmo the inputs list do not returning NAs
}
