% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_sim_helpers.R
\name{make_tax_brackets}
\alias{make_tax_brackets}
\title{Create a data frame with the tax brakes}
\usage{
make_tax_brackets(thresholds, rates)
}
\description{
Requires two vectors of equal length. One with the upper tax thresholds and
another one with the the tax rates
}
