% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_sim_helpers.R
\name{match_tax_brackets}
\alias{match_tax_brackets}
\title{Apply tax brackets to the data and impute new tax variable}
\usage{
match_tax_brackets(
  dta,
  base_var,
  rules,
  new_var,
  calc_drop = TRUE,
  low_thld_rule = `>`,
  up_thld_rule = `<=`
)
}
\arguments{
\item{dta}{original data set to compute tax levels}

\item{base_var}{string variable name that should be used used as income}

\item{rules}{data frame with tax brackets rules returned with `make_tax_brackets`}

\item{new_var}{string of the new variable to create based on the formula
cum_tax + (base_var - lower) * rate}

\item{calc_drop}{logical TRUE as default to compute the tax level and drop
all auxiliary variables}

\item{low_thld_rule}{rule for comparing income to the lower threshold.
usually `>` (default) or `>`.}

\item{up_thld_rule}{rule for comparing income to the upper threshold.
usually `<=`  (default) or `<`.}
}
\description{
Apply tax brackets to the data and impute new tax variable
}
