% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{microsim_dta}
\alias{microsim_dta}
\title{Exemplary microsimulation results data}
\format{
## `microsim_dta`
A data frame with 7,240 rows and 60 columns:
\describe{
  \item{country}{Country name}
  \item{iso2, iso3}{2 & 3 letter ISO country codes}
  \item{year}{Year}
  ...
}
}
\source{
data-raw/DATASETS.R
}
\usage{
microsim_dta
}
\description{
This is a fictitious data set meant to show what the ideal misro-simulation
  results data should look like.
}
\keyword{datasets}
