% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ceq_progress.R
\name{mod_ceq_progress}
\alias{mod_ceq_progress}
\alias{fct_make_ceq_progress}
\alias{fct_big_step_ceq_progress}
\alias{fct_samll_step_ceq_progress}
\alias{fct_close_ceq_progress}
\title{CEQ progress module that initializes the progress object}
\usage{
mod_ceq_progress(id, run = reactive(NULL), prog_length = 4)

fct_make_ceq_progress(session, prog_length = 4)

fct_big_step_ceq_progress(prog, message = NULL, detail = NULL, ...)

fct_samll_step_ceq_progress(
  prog,
  n_small = 4,
  message = NULL,
  detail = NULL,
  ...
)

fct_close_ceq_progress(prog, title = "Simulation is completed", ...)
}
\description{
CEQ progress module that initializes the progress object
}
\section{Functions}{
\itemize{
\item \code{fct_make_ceq_progress()}: Initialize progress.

\item \code{fct_big_step_ceq_progress()}: big step progress

\item \code{fct_samll_step_ceq_progress()}: small step progress

\item \code{fct_close_ceq_progress()}: small step progress

}}
