% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_generic_run_postsim.R
\name{mod_generic_run_postsim_server}
\alias{mod_generic_run_postsim_server}
\title{Generic post-simulation runner server}
\usage{
mod_generic_run_postsim_server(
  id,
  sim_res = reactive(NULL),
  ceq_progress = NULL,
  fn_ceq_postsim = function(x, progress_tick = function() {
 }) {
     progress_tick()
  
      x
 },
  ...
)
}
\arguments{
\item{sim_res}{reactive containing results of the `mod_generic_run_sim_server`}

\item{ceq_progress}{is reactive tha contains progress function created
with the `fct_make_ceq_progress`}

\item{fn_ceq_postsim}{a function that performs post-simulation
calculations that is relevant to perform at this stage. This, for
example could be aggregating deciles or creating a bulk of default plots.
it accepts two arguments `sim_res()` and `progress_tick`, which is the
function that ticks progress.}
}
\value{
a `reactiveValues` object that contains a list of data sets with
  post simulation results that are usually can be directly plotted.
}
\description{
Generic post-simulation runner server
}
