% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_inputs_tabs.R
\name{mod_render_inp_tabs_srv}
\alias{mod_render_inp_tabs_srv}
\title{Module for generating and re-generating inputs UI on the server}
\usage{
mod_render_inp_tabs_srv(
  id,
  switches = reactive(get_test_tab_switches()),
  tabs = reactive(get_test_tabs(id)),
  summary_tab = reactive(mtcars),
  header = reactive(NULL),
  tab_header = reactive(NULL),
  tab_footer = reactive(NULL),
  ...
)
}
\arguments{
\item{switches}{reactive that returns a list with key information needed
for tab switches to be rendered:}
}
\description{
Module for generating and re-generating inputs UI on the server
}
