% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_generic_run_sim.R
\name{test_mod_run_sim}
\alias{test_mod_run_sim}
\title{Simple app for testing module `mod_generic_run_sim_server`
  with the user-defined logic}
\usage{
test_mod_run_sim(
  sim_inp_path = NULL,
  presim_path = NULL,
  raw_str_path = NULL,
  id = NULL,
  fn_sim_srvr = make_run_sim_server(),
  inps = reactive(list(policy0 = list(policy_choices = list(a = 1, b = 2), policy_as_base
    = FALSE, policy_name = "Test 1"), policy1 = list(policy_choices = list(a = 3, b = 4),
    policy_as_base = FALSE, policy_name = "Test 2"))),
  all_inps = reactive(list(a = 1, b = 2, c = 3, d = 3)),
  presim = reactive(list(baseline = tibble(baseline = 1:10), presim_1 = tibble(presim =
    c(5:9))))
)
}
\description{
Simple app for testing module `mod_generic_run_sim_server`
  with the user-defined logic
}
