% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_incidences.R
\name{test_results_mod}
\alias{test_results_mod}
\title{Tests a simple results page module providing it with the sim_res}
\usage{
test_results_mod(sim_res = NULL, ui_side, server_side, id = "testid")
}
\description{
Tests a simple results page module providing it with the sim_res
}
