% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_wb_colours.R
\docType{data}
\name{wb_cols}
\alias{wb_cols}
\alias{wb_palettes}
\alias{wb_pal}
\alias{scale_color_wb}
\alias{scale_fill_wb}
\title{Function to extract wb colors as hex codes}
\format{
An object of class \code{list} of length 6.
}
\usage{
wb_cols(...)

wb_palettes

wb_pal(palette = "main", reverse = FALSE, ...)

scale_color_wb(palette = "main2", discrete = TRUE, reverse = FALSE, ...)

scale_fill_wb(palette = "main2", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradientn(), used respectively when discrete is TRUE or FALSE}

\item{palette}{Character name of palette in wb_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}
}
\description{
Function to extract wb colors as hex codes

Return function to interpolate a wb color palette

Color scale constructor for wb colors

Fill scale constructor for wb colors
}
\examples{

\dontrun{
wb_cols()
wb_cols("red")
wb_cols("red", "blue")
wb_cols("blue", "red")

library(ggplot2)
ggplot(mtcars, aes(hp, mpg)) +
  geom_point(color = wb_cols("red"),
             size = 4, alpha = .8)
}

\dontrun{
wb_pal("main2")
wb_pal("neutral_all")(10)
}


\dontrun{
library(ggplot2)
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +
  geom_point(size = 4) +
  scale_color_wb(palette = "main2")

ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Sepal.Length)) +
  geom_point(size = 4, alpha = .6) +
  scale_color_wb(discrete = FALSE, palette = "neutral_warm")


ggplot(mpg, aes(manufacturer, fill = manufacturer)) +
  geom_bar() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  scale_fill_wb(palette = "main3", guide = "none")
}

}
\keyword{datasets}
