% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_input_ui.R
\name{load_input_xlsx}
\alias{load_input_xlsx}
\alias{inp_str_test}
\alias{gen_inp_str_front}
\alias{gen_inp_str}
\alias{test_gen_inp_front_simple}
\alias{test_gen_inp_front_tabs}
\alias{test_gen_inp_front_tabs_file}
\title{Loads the inputs structure data into R  from an excel file.}
\usage{
load_input_xlsx(path)

inp_str_test(
  inp_raw_str,
  fn_inp_str = gen_inp_str,
  fn_inp_ui = gen_tabinp_ui,
  ns = NS(NULL)
)

gen_inp_str_front(inp_table_str = NULL, ...)

gen_inp_str(inp_raw_str, n_choices, inp_table_str = NULL, ns = NS(NULL), ...)

test_gen_inp_front_simple(
  inp_raw_str,
  inp_tab_str = NULL,
  inp_table_str = NULL,
  n_choices = 2
)

test_gen_inp_front_tabs(
  inp_raw_str,
  inp_tab_str = NULL,
  inp_table_str = NULL,
  n_choices = 2
)

test_gen_inp_front_tabs_file(path)
}
\arguments{
\item{path}{path to a pre-defined excel file}

\item{inp_raw_str}{inputs structure as returned by `load_input_xlsx()`}

\item{fn_inp_str, fn_inp_ui, ns}{internal supporting functions.}

\item{ns}{namestapce function `NS(NULL)` by default}

\item{inp_table_str}{structure of tables under which inputs are organized (NULL by default)}

\item{...}{passed to functions within}

\item{n_choices}{number of policy choices to generate}

\item{inp_raw_str, inp_tab_str, inp_table_str, n_choices}{input layout options}
}
\value{
a data frame with the detailed inputs structure.
}
\description{
Loads the inputs structure data into R  from an excel file.
}
\section{Functions}{
\itemize{
\item \code{inp_str_test()}: Test if RAW inputs structure generate a set of UIs

\item \code{gen_inp_str_front()}: Wrapper around `gen_inp_str()` used for providing tables structure

\item \code{gen_inp_str()}: Generate a data frame of the inputs UI used in creating the inputs page UI

\item \code{test_gen_inp_front_simple()}: Test input page content generation in a single tab

\item \code{test_gen_inp_front_tabs()}: Test input tabs content generation process by providing basic data.

\item \code{test_gen_inp_front_tabs_file()}: Test input tabs content generation process by providing basic data.

}}
\examples{
\dontrun{
system.file(
  "examples",
  "ceq_example_simple",
  "data-app",
  "simple-inputs-structure.xlsx",
  package = "devCEQ"
) |> load_input_xlsx()
}


\dontrun{
system.file(
  "examples",
  "ceq_example_simple",
  "data-app",
  "simple-inputs-structure.xlsx",
  package = "devCEQ"
) |>
  load_input_xlsx() |>
  inp_str_test()
}

\dontrun{
a <- system.file(
    "examples",
    "ceq_example_simple",
    "data-app",
    "simple-inputs-structure.xlsx",
    package = "devCEQ"
  ) |> load_input_xlsx()
a |> test_gen_inp_front_simple()
}

\dontrun{
a <- system.file(
    "examples",
    "ceq_example_simple",
    "data-app",
    "simple-inputs-structure.xlsx",
    package = "devCEQ"
  ) |> load_input_xlsx()
a |> test_gen_inp_front_tabs()
}
\dontrun{
system.file(
    "examples",
    "ceq_example_simple",
    "data-app",
    "simple-inputs-structure.xlsx",
    package = "devCEQ"
  ) |>
  test_gen_inp_front_tabs_file()
}
}
